build(deps): pin dependencies and unify framework patches
- Pin `camouchat-whatsapp`, `python-dotenv`, `pytz`, and `art` to specific versions in `pyproject.toml`.
- Add `qrcode` to core dependencies.
- Migrate logging filters and formatters from `app.py` to a centralized `patches.py`.
- Replace `patch_status_get.py` with a unified `apply_all_patches()` entry point.
- Clean up `app.py` by removing inline logging logic and consolidating patch application.
