+++ pyproject.toml
@@ -5,12 +5,13 @@ description = "WhatsApp Status Checker using CamouChat"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
-    "camouchat-whatsapp>=0.7.3",
-    "python-dotenv>=1.1.1",
-    "pytz>=2026.2",
-    "art>=6.1.1",
-    "callmebot",
-    "python-dotenv"
+    "camouchat-whatsapp==0.7.3",
+    "python-dotenv==1.2.2",
+    "pytz==2026.2",
+    "art==6.5",
+    "callmebot>=1.0.2",
+    "python-dotenv==1.2.2",
+    "qrcode==7.4.2"
 ]
 
 [tool.setuptools.packages.find]
